requirements.txt
flask
onnxruntime
opencv-python
numpy
pillow
torchvision
requests

